requirements.txt
alembic==1.3.3
bcrypt==3.1.7
blinker==1.4
certifi==2019.11.28
cffi==1.13.2
Click==7.0
Flask==1.1.1
Flask-Bcrypt==0.7.1
Flask-Login==0.4.1
Flask-Mail==0.9.1
flask-marshmallow==0.10.1
Flask-Migrate==2.5.2
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.2
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.1
Mako==1.1.1
MarkupSafe==1.1.1
marshmallow==3.4.0
marshmallow-sqlalchemy==0.21.0
psycopg2-binary==2.8.4
pycparser==2.19
python-dateutil==2.8.1
python-dotenv==0.10.5
python-editor==1.0.4
six==1.14.0
SQLAlchemy==1.3.13
Werkzeug==0.16.1
WTForms==2.2.1
